+++ robe-common/src/main/java/io/robe/common/dto/BooleanHolder.java
@@ -1,0 +1,19 @@
+package io.robe.common.dto;
+
+/**
+ * Created by kamilbukum on 30/01/2017.
+ */
+public class BooleanHolder {
+    private boolean value;
+    public BooleanHolder(boolean value){
+        this.value = value;
+    }
+
+    public void set(boolean value) {
+        this.value = value;
+    }
+
+    public boolean is() {
+        return value;
+    }
+}