+++ robe-hibernate/src/main/java/io/robe/hibernate/criteria/api/cache/DTOMeta.java
@@ -1,7 +1,0 @@
-package io.robe.hibernate.criteria.api.cache;
-
-/**
- * Created by kamilbukum on 23/01/2017.
- */
-public class DTOMeta {
-}