+++ robe-hibernate/src/main/java/io/robe/hibernate/criteria/query/search/SearchFilter.java
@@ -1,0 +1,13 @@
+package io.robe.hibernate.criteria.query.search;
+
+/**
+ * Created by kamilbukum on 27/01/2017.
+ */
+public class SearchFilter {
+    private String name;
+    private String op;
+    private String value;
+    public SearchFilter(String name, String op, String value) {
+
+    }
+}