+++ robe-hibernate/src/test/java/io/robe/hibernate/criteria/api/criteria/CriteriaTest.java
@@ -96,9 +96,9 @@ public class CriteriaTest extends HqlCriteriaTestTools {
         Criteria<Map<String, Object>> mapCriteria = Criteria
                 .createCriteria(User.class, new TransformerImpl<>(session, Criteria.MAP_CLASS))
                 .setProjection(Projections.projectionList().add(
-                        Projections.groupProperty("name"),
+                        Projections.alias( Projections.max("failCount"), "max_failCount"),
                         "name"
-                ));
+                )).addOrder(Order.asc("max_failCount"));
         mapCriteria
                 .createJoin(Role.class, "roleOid")
                 .setProjection(Projections.alias(